squash warnings

+++ engine/class_modules/paladin/sc_paladin.hpp
@@ -751,8 +751,8 @@ public:
   {
     bool fake_sov                         = true;
     double proc_chance_ret_aura_sera      = 0.10;
-    double min_dg_heal_targets            = 1.0;
-    double max_dg_heal_targets            = 5.0;
+    int min_dg_heal_targets               = 1;
+    int max_dg_heal_targets               = 5;
     bool sacred_weapon_prefer_new_targets = false;
   } options;
   player_t* beacon_target;
